+++ src/Dapr.Actors.AspNetCore/ActorsServiceCollectionExtensions.cs
@@ -47,15 +47,14 @@ namespace Microsoft.Extensions.DependencyInjection
             {   
                 var options = s.GetRequiredService<IOptions<ActorRuntimeOptions>>().Value;
 
-                //Replace the HttpEndpoint with an endpoint prioritizing IConfiguration
                 var configuration = s.GetService<IConfiguration>();
-                options.HttpEndpoint = options.HttpEndpoint != "http://127.0.0.1:3500"
-                    ? GetHttpEndpoint(configuration)
-                    : options.HttpEndpoint;
-                options.DaprApiToken = string.IsNullOrWhiteSpace(options.DaprApiToken)
-                    ? GetApiToken(configuration)
-                    : options.DaprApiToken;
-
+                options.DaprApiToken = !string.IsNullOrWhiteSpace(options.DaprApiToken)
+                    ? options.DaprApiToken
+                    : DaprDefaults.GetDefaultDaprApiToken(configuration);
+                options.HttpEndpoint = !string.IsNullOrWhiteSpace(options.HttpEndpoint)
+                    ? options.HttpEndpoint
+                    : DaprDefaults.GetDefaultHttpEndpoint();
+                
                 var loggerFactory = s.GetRequiredService<ILoggerFactory>();
                 var activatorFactory = s.GetRequiredService<ActorActivatorFactory>();
                 var proxyFactory = s.GetRequiredService<IActorProxyFactory>();
@@ -68,12 +67,12 @@ namespace Microsoft.Extensions.DependencyInjection
 
                 //Replace the HttpEndpoint with an endpoint prioritizing IConfiguration
                 var configuration = s.GetService<IConfiguration>();
-                options.HttpEndpoint = options.HttpEndpoint != "http://127.0.0.1:3500"
-                    ? GetHttpEndpoint(configuration)
-                    : options.HttpEndpoint;
-                options.DaprApiToken = string.IsNullOrWhiteSpace(options.DaprApiToken)
-                    ? GetApiToken(configuration)
-                    : options.DaprApiToken;
+                options.DaprApiToken = !string.IsNullOrWhiteSpace(options.DaprApiToken)
+                    ? options.DaprApiToken
+                    : DaprDefaults.GetDefaultDaprApiToken(configuration);
+                options.HttpEndpoint = !string.IsNullOrWhiteSpace(options.HttpEndpoint)
+                    ? options.HttpEndpoint
+                    : DaprDefaults.GetDefaultHttpEndpoint();
 
                 var factory = new ActorProxyFactory() 
                 { 
@@ -92,89 +91,6 @@ namespace Microsoft.Extensions.DependencyInjection
             {
                 services.Configure<ActorRuntimeOptions>(configure);
             }
-        }
-
-        /// <summary>
-        /// Retrieves the Dapr API token using a failover approach starting with an optional <see cref="IConfiguration"/>
-        /// instance, then trying to pull from the well-known environment variable name and then opting for an empty string
-        /// as a default value.
-        /// </summary>
-        /// <returns>The Dapr API token.</returns>
-        private static string GetApiToken(IConfiguration? configuration) => GetResourceValue(configuration, DaprDefaults.DaprApiTokenName);
-
-        /// <summary>
-        /// Builds the Dapr gRPC endpoint using the value from the IConfiguration, if available, then falling back
-        /// to the value in the environment variable(s) and finally otherwise using the default value (an empty string).
-        /// </summary>
-        /// <remarks>
-        /// Marked as internal for testing purposes.
-        /// </remarks>
-        /// <param name="configuration">An injected instance of the <see cref="IConfiguration"/>.</param>
-        /// <returns>The built gRPC endpoint.</returns>
-        private static string GetHttpEndpoint(IConfiguration? configuration)
-        {
-            //Prioritize pulling from IConfiguration with a fallback from pulling from the environment variable directly
-            var httpEndpoint = GetResourceValue(configuration, DaprDefaults.DaprHttpEndpointName);
-            var httpPort = GetResourceValue(configuration, DaprDefaults.DaprHttpPortName);
-            int? parsedGrpcPort = string.IsNullOrWhiteSpace(httpPort) ? null : int.Parse(httpPort);
-
-            var endpoint = BuildEndpoint(httpEndpoint, parsedGrpcPort);
-            return string.IsNullOrWhiteSpace(endpoint) ? $"http://localhost:{DaprDefaults.DefaultHttpPort}/" : endpoint;
-        }
-
-        /// <summary>
-        /// Retrieves the specified value prioritizing pulling it from <see cref="IConfiguration"/>, falling back
-        /// to an environment variable, and using an empty string as a default.
-        /// </summary>
-        /// <param name="configuration">An instance of an <see cref="IConfiguration"/>.</param>
-        /// <param name="name">The name of the value to retrieve.</param>
-        /// <returns>The value of the resource.</returns>
-        private static string GetResourceValue(IConfiguration? configuration, string name)
-        {
-            //Attempt to retrieve first from the configuration
-            var configurationValue = configuration?.GetValue<string?>(name);
-            if (configurationValue is not null)
-                return configurationValue;
-
-            //Fall back to the environment variable with the same name or default to an empty string
-            var envVar = Environment.GetEnvironmentVariable(name);
-            return envVar ?? string.Empty;
-        }
-
-        /// <summary>
-        /// Builds the endpoint provided an optional endpoint and optional port.
-        /// </summary>
-        /// <remarks>
-        /// Marked as internal for testing purposes.
-        /// </remarks>
-        /// <param name="endpoint">The endpoint.</param>
-        /// <param name="endpointPort">The port</param>
-        /// <returns>A constructed endpoint value.</returns>
-        internal static string BuildEndpoint(string? endpoint, int? endpointPort)
-        {
-            if (string.IsNullOrWhiteSpace(endpoint) && endpointPort is null)
-                return string.Empty;
-
-            var endpointBuilder = new UriBuilder();
-            if (!string.IsNullOrWhiteSpace(endpoint))
-            {
-                //Extract the scheme, host and port from the endpoint
-                var uri = new Uri(endpoint);
-                endpointBuilder.Scheme = uri.Scheme;
-                endpointBuilder.Host = uri.Host;
-                endpointBuilder.Port = uri.Port;
-
-                //Update the port if provided separately
-                if (endpointPort is not null)
-                    endpointBuilder.Port = (int)endpointPort;
-            }
-            else if (string.IsNullOrWhiteSpace(endpoint) && endpointPort is not null)
-            {
-                endpointBuilder.Host = "localhost";
-                endpointBuilder.Port = (int)endpointPort;
-            }
-
-            return endpointBuilder.ToString();
         }
     }
 }
