+++ src/Dapr.AspNetCore/DaprAuthenticationOptions.cs
@@ -11,7 +11,6 @@
 // limitations under the License.
 // ------------------------------------------------------------------------
 
-using System;
 using Microsoft.AspNetCore.Authentication;
 
 namespace Dapr.AspNetCore
@@ -29,6 +28,6 @@ namespace Dapr.AspNetCore
         /// Gets or sets the App API token.
         /// By default, the token will be read from the APP_API_TOKEN environment variable.
         /// </summary>
-        public string Token { get; set; } = DaprDefaults.GetDefaultAppApiToken();
+        public string Token { get; set; } = DaprDefaults.GetDefaultAppApiToken(null);
     }
 }
