+++ src/Dapr.AspNetCore/DaprServiceCollectionExtensions.cs
@@ -42,18 +42,14 @@ public static class DaprServiceCollectionExtensions
 
             var configuration = serviceProvider.GetService<IConfiguration>();
 
-            //Set the HTTP endpoint, if provided
-            var httpEndpoint = GetHttpEndpoint(configuration);
-            if (!string.IsNullOrWhiteSpace(httpEndpoint))
-                builder.UseHttpEndpoint(httpEndpoint);
+            //Set the HTTP endpoint, if provided, else use the default endpoint
+            builder.UseHttpEndpoint(DaprDefaults.GetDefaultHttpEndpoint(configuration));
 
             //Set the gRPC endpoint, if provided
-            var grpcEndpoint = GetGrpcEndpoint(configuration);
-            if (!string.IsNullOrWhiteSpace(grpcEndpoint))
-                builder.UseGrpcEndpoint(grpcEndpoint);
-
+            builder.UseGrpcEndpoint(DaprDefaults.GetDefaultGrpcEndpoint(configuration));
+            
             //Set the API token, if provided
-            var apiToken = GetApiToken(configuration);
+            var apiToken = DaprDefaults.GetDefaultDaprApiToken(configuration);
             if (!string.IsNullOrWhiteSpace(apiToken))
                 builder.UseDaprApiToken(apiToken);
 
@@ -80,18 +76,14 @@ public static class DaprServiceCollectionExtensions
 
             var configuration = serviceProvider.GetService<IConfiguration>();
 
-            //Set the HTTP endpoint, if provided
-            var httpEndpoint = GetHttpEndpoint(configuration);
-            if (!string.IsNullOrWhiteSpace(httpEndpoint))
-                builder.UseHttpEndpoint(httpEndpoint);
+            //Set the HTTP endpoint, if provided, else use the default endpoint
+            builder.UseHttpEndpoint(DaprDefaults.GetDefaultHttpEndpoint(configuration));
 
             //Set the gRPC endpoint, if provided
-            var grpcEndpoint = GetGrpcEndpoint(configuration);
-            if (!string.IsNullOrWhiteSpace(grpcEndpoint))
-                builder.UseGrpcEndpoint(grpcEndpoint);
-
+            builder.UseGrpcEndpoint(DaprDefaults.GetDefaultGrpcEndpoint(configuration));
+            
             //Set the API token, if provided
-            var apiToken = GetApiToken(configuration);
+            var apiToken = DaprDefaults.GetDefaultDaprApiToken(configuration);
             if (!string.IsNullOrWhiteSpace(apiToken))
                 builder.UseDaprApiToken(apiToken);
 
@@ -99,122 +91,5 @@ public static class DaprServiceCollectionExtensions
 
             return builder.Build();
         });
-    }
-
-    /// <summary>
-    /// Builds the Dapr HTTP endpoint using the value from the IConfiguration, if available, then falling back
-    /// to the value in the environment variable(s) and finally otherwise using the default value (an empty string).
-    /// </summary>
-    /// <remarks>
-    /// Marked as internal for testing purposes.
-    /// </remarks>
-    /// <param name="configuration">An injected instance of the <see cref="IConfiguration"/>.</param>
-    /// <returns>The built HTTP endpoint.</returns>
-    internal static string GetHttpEndpoint(IConfiguration? configuration)
-    {
-        //Prioritize pulling from IConfiguration with a fallback of pulling from the environment variable directly
-        var httpEndpoint = GetResourceValue(configuration, DaprDefaults.DaprHttpEndpointName);
-        var httpPort = GetResourceValue(configuration, DaprDefaults.DaprHttpPortName);
-        int? parsedHttpPort = int.TryParse(httpPort, out var port) ? port : null;
-
-        var endpoint = BuildEndpoint(httpEndpoint, parsedHttpPort);
-        return string.IsNullOrWhiteSpace(endpoint)
-            ? $"http://{DaprDefaults.DaprHostName}:{DaprDefaults.DefaultHttpPort}/"
-            : endpoint;
-    }
-
-    /// <summary>
-    /// Builds the Dapr gRPC endpoint using the value from the IConfiguration, if available, then falling back
-    /// to the value in the environment variable(s) and finally otherwise using the default value (an empty string).
-    /// </summary>
-    /// <remarks>
-    /// Marked as internal for testing purposes.
-    /// </remarks>
-    /// <param name="configuration">An injected instance of the <see cref="IConfiguration"/>.</param>
-    /// <returns>The built gRPC endpoint.</returns>
-    internal static string GetGrpcEndpoint(IConfiguration? configuration)
-    {
-        //Prioritize pulling from IConfiguration with a fallback from pulling from the environment variable directly
-        var grpcEndpoint = GetResourceValue(configuration, DaprDefaults.DaprGrpcEndpointName);
-        var grpcPort = GetResourceValue(configuration, DaprDefaults.DaprGrpcPortName);
-        int? parsedGrpcPort = int.TryParse(grpcPort, out var port) ? port : null;
-
-        var endpoint = BuildEndpoint(grpcEndpoint, parsedGrpcPort);
-        return string.IsNullOrWhiteSpace(endpoint)
-            ? $"http://{DaprDefaults.DaprHostName}:{DaprDefaults.DefaultGrpcPort}/"
-            : endpoint;
-    }
-
-    /// <summary>
-    /// Retrieves the Dapr API token first from the <see cref="IConfiguration"/>, if available, then falling back
-    /// to the value in the environment variable(s) directly, then finally otherwise using the default value (an
-    /// empty string).
-    /// </summary>
-    /// <remarks>
-    /// Marked as internal for testing purposes.
-    /// </remarks>
-    /// <param name="configuration">An injected instance of the <see cref="IConfiguration"/>.</param>
-    /// <returns>The Dapr API token.</returns>
-    internal static string GetApiToken(IConfiguration? configuration)
-    {
-        //Prioritize pulling from IConfiguration with a fallback of pulling from the environment variable directly
-        return GetResourceValue(configuration, DaprDefaults.DaprApiTokenName);
-    }
-
-    /// <summary>
-    /// Retrieves the specified value prioritizing pulling it from <see cref="IConfiguration"/>, falling back
-    /// to an environment variable, and using an empty string as a default.
-    /// </summary>
-    /// <param name="configuration">An instance of an <see cref="IConfiguration"/>.</param>
-    /// <param name="name">The name of the value to retrieve.</param>
-    /// <returns>The value of the resource.</returns>
-    private static string GetResourceValue(IConfiguration? configuration, string name)
-    {
-        //Attempt to retrieve first from the configuration
-        var configurationValue = configuration?.GetValue<string?>(name);
-        if (configurationValue is not null)
-            return configurationValue;
-
-        //Fall back to the environment variable with the same name or default to an empty string
-        var envVar = Environment.GetEnvironmentVariable(name);
-        return envVar ?? string.Empty;
-    }
-
-    /// <summary>
-    /// Builds the endpoint provided an optional endpoint and optional port.
-    /// </summary>
-    /// <remarks>
-    /// Marked as internal for testing purposes.
-    /// </remarks>
-    /// <param name="endpoint">The endpoint.</param>
-    /// <param name="endpointPort">The port</param>
-    /// <returns>A constructed endpoint value.</returns>
-    internal static string BuildEndpoint(string? endpoint, int? endpointPort)
-    {
-        if (string.IsNullOrWhiteSpace(endpoint) && endpointPort is null)
-            return string.Empty;
-        
-        // Per the proposal at https://github.com/artursouza/proposals/blob/cd811136d0af0aade52ef297a84c3050f3243ae8/0008-S-sidecar-endpoint-tls.md#design this will
-        // favor whatever value is provided in the endpoint value (and if the port isn't provided, it will be inferred from the protocol).
-        
-        // The endpoint port will only be evaluated if the endpoint is not provided. While the proposal calls for a value of "127.0.0.1", it does accept its
-        // equivalent, e.g. "localhost", and because of the issue detailed at https://github.com/dapr/dotnet-sdk/issues/1032
-        
-        var endpointBuilder = new UriBuilder();
-        if (!string.IsNullOrWhiteSpace(endpoint)) //Ignore the endpoint port argument if the endpoint is provided
-        {
-            //Extract the scheme, host and port from the endpoint
-            var uri = new Uri(endpoint);
-            endpointBuilder.Scheme = uri.Scheme;
-            endpointBuilder.Host = uri.Host;
-            endpointBuilder.Port = uri.Port;
-        }
-        else if (string.IsNullOrWhiteSpace(endpoint) && endpointPort is not null)
-        {
-            endpointBuilder.Host = "localhost";
-            endpointBuilder.Port = (int)endpointPort;
-        }
-
-        return endpointBuilder.ToString();
     }
 }
