+++ src/Dapr.Common/Dapr.Common.csproj
@@ -9,6 +9,7 @@
     <ItemGroup>
         <PackageReference Include="Grpc.Net.Client" />
         <PackageReference Include="Microsoft.Extensions.Http" />
+        <PackageReference Include="Microsoft.Extensions.Configuration" />
     </ItemGroup>
 
 </Project>
