+++ src/Dapr.Common/DaprDefaults.cs
@@ -11,6 +11,8 @@
 // limitations under the License.
 // ------------------------------------------------------------------------
 
+using Microsoft.Extensions.Configuration;
+
 namespace Dapr
 {
     internal static class DaprDefaults
@@ -27,88 +29,104 @@ namespace Dapr
         public const string DaprGrpcEndpointName = "DAPR_GRPC_ENDPOINT";
         public const string DaprGrpcPortName = "DAPR_GRPC_PORT";
 
-        public const string DaprHostName = "localhost";
+        public const string DefaultDaprScheme = "http";
+        public const string DefaultDaprHost = "localhost";
         public const int DefaultHttpPort = 3500;
         public const int DefaultGrpcPort = 50001;
 
         /// <summary>
         /// Get the value of environment variable DAPR_API_TOKEN
         /// </summary>
+        /// <param name="configuration">The optional <see cref="IConfiguration"/> to pull the value from.</param>
         /// <returns>The value of environment variable DAPR_API_TOKEN</returns>
-        public static string GetDefaultDaprApiToken()
-        {
-            // Lazy-init is safe because this is just populating the default
-            // We don't plan to support the case where the user changes environment variables
-            // for a running process.
-            if (string.IsNullOrEmpty(daprApiToken))
-            {
-                // Treat empty the same as null since it's an environment variable
-                var value = Environment.GetEnvironmentVariable(DaprApiTokenName);
-                daprApiToken = string.IsNullOrEmpty(value) ? string.Empty : value;
-            }
-
-            return daprApiToken;
-        }
+        public static string GetDefaultDaprApiToken(IConfiguration? configuration) =>
+            GetResourceValue(configuration, DaprApiTokenName) ?? string.Empty;
 
         /// <summary>
         /// Get the value of environment variable APP_API_TOKEN
         /// </summary>
+        /// <param name="configuration">The optional <see cref="IConfiguration"/> to pull the value from.</param>
         /// <returns>The value of environment variable APP_API_TOKEN</returns>
-        public static string GetDefaultAppApiToken()
-        {
-            if (string.IsNullOrEmpty(appApiToken))
-            {
-                var value = Environment.GetEnvironmentVariable(AppApiTokenName);
-                appApiToken = string.IsNullOrEmpty(value) ? string.Empty : value;
-            }
-
-            return appApiToken;
-        }
+        public static string GetDefaultAppApiToken(IConfiguration? configuration) =>
+            GetResourceValue(configuration, AppApiTokenName) ?? string.Empty;
 
         /// <summary>
         /// Get the value of HTTP endpoint based off environment variables
         /// </summary>
+        /// <param name="configuration">The optional <see cref="IConfiguration"/> to pull the value from.</param>
         /// <returns>The value of HTTP endpoint based off environment variables</returns>
-        public static string GetDefaultHttpEndpoint()
+        public static string GetDefaultHttpEndpoint(IConfiguration? configuration = null)
         {
-            if (string.IsNullOrEmpty(httpEndpoint))
-            {
-                var endpoint = Environment.GetEnvironmentVariable(DaprHttpEndpointName);
-                if (!string.IsNullOrEmpty(endpoint))
-                {
-                    httpEndpoint = endpoint;
-                    return httpEndpoint;
-                }
+            //Prioritize pulling from the IConfiguration and fallback to the environment variable if not populated
+            var endpoint = GetResourceValue(configuration, DaprHttpEndpointName);
+            var port = GetResourceValue(configuration, DaprHttpPortName);
+            
+            //Use the default HTTP port if we're unable to retrieve/parse the provided port
+            int? parsedGrpcPort = string.IsNullOrWhiteSpace(port) ? DefaultHttpPort : int.Parse(port);
 
-                var port = Environment.GetEnvironmentVariable(DaprHttpPortName);
-                port = string.IsNullOrEmpty(port) ? $"{DefaultHttpPort}" : port;
-                httpEndpoint = $"http://127.0.0.1:{port}";
-            }
-
-            return httpEndpoint;
+            return BuildEndpoint(endpoint, parsedGrpcPort.Value);
         }
 
         /// <summary>
         /// Get the value of gRPC endpoint based off environment variables
         /// </summary>
+        /// <param name="configuration">The optional <see cref="IConfiguration"/> to pull the value from.</param>
         /// <returns>The value of gRPC endpoint based off environment variables</returns>
-        public static string GetDefaultGrpcEndpoint()
+        public static string GetDefaultGrpcEndpoint(IConfiguration? configuration = null)
         {
-            if (string.IsNullOrEmpty(grpcEndpoint))
-            {
-                var endpoint = Environment.GetEnvironmentVariable(DaprGrpcEndpointName);
-                if (!string.IsNullOrEmpty(endpoint))
-                {
-                    grpcEndpoint = endpoint;
-                    return grpcEndpoint;
-                }
+            //Prioritize pulling from the IConfiguration and fallback to the environment variable if not populated
+            var endpoint = GetResourceValue(configuration, DaprGrpcEndpointName);
+            var port = GetResourceValue(configuration, DaprGrpcPortName);
+            
+            //Use the default gRPC port if we're unable to retrieve/parse the provided port
+            int? parsedGrpcPort = string.IsNullOrWhiteSpace(port) ? DefaultGrpcPort : int.Parse(port);
 
-                var port = Environment.GetEnvironmentVariable(DaprGrpcPortName);
-                port = string.IsNullOrEmpty(port) ? $"{DefaultGrpcPort}" : port;
-                grpcEndpoint = $"http://127.0.0.1:{port}";
+            return BuildEndpoint(endpoint, parsedGrpcPort.Value);
+        }
+
+        /// <summary>
+        /// Builds the Dapr endpoint.
+        /// </summary>
+        /// <param name="endpoint">The endpoint value.</param>
+        /// <param name="endpointPort">The endpoint port value, whether pulled from configuration/envvar or the default.</param>
+        /// <returns>A constructed endpoint value.</returns>
+        private static string BuildEndpoint(string? endpoint, int endpointPort)
+        {
+            var endpointBuilder = new UriBuilder { Scheme = DefaultDaprScheme, Host = DefaultDaprHost }; //Port depends on endpoint
+
+            if (!string.IsNullOrWhiteSpace(endpoint)) //If the endpoint is set, it doesn't matter if the port is
+            {
+                //Extract the scheme, host and port from the endpoint and replace defaults
+                var uri = new Uri(endpoint);
+                endpointBuilder.Scheme = uri.Scheme;
+                endpointBuilder.Host = uri.Host;
+                endpointBuilder.Port = uri.Port;
+            }
+            else
+            {
+                //Should only set the port if the endpoint isn't populated
+                endpointBuilder.Port = endpointPort;
             }
 
-            return grpcEndpoint;
+            return endpointBuilder.ToString();
+        }
+        
+        /// <summary>
+        /// Retrieves the specified value prioritizing pulling it from <see cref="IConfiguration"/>, falling back
+        /// to an environment variable, and using an empty string as a default.
+        /// </summary>
+        /// <param name="configuration">An instance of an <see cref="IConfiguration"/>.</param>
+        /// <param name="name">The name of the value to retrieve.</param>
+        /// <returns>The value of the resource.</returns>
+        private static string? GetResourceValue(IConfiguration? configuration, string name)
+        {
+            //Attempt to retrieve first from the configuration
+            var configurationValue = configuration?[name];
+            if (configurationValue is not null)
+                return configurationValue;
+
+            //Fall back to the environment variable with the same name or default to an empty string
+            return Environment.GetEnvironmentVariable(name);
         }
     }
 }
